Implement parameters checking in EFGenericRepository CRUD methods
optimizing linq query, considing navigation property with eager loading

Refctor SaveChanges calls and exception catching (consideration of async not yet covered)

+++ CareerCloud.EntityFrameworkDataAccess/EFGenericRepository.cs
@@ -3,10 +3,16 @@ using CareerCloud.DataAccessLayer;
 using Microsoft.EntityFrameworkCore;
 
 namespace CareerCloud.EntityFrameworkDataAccess;
+
+/*  Future considerations:
+*   Async methods
+*   Centralized logging of certain exceptions or stats
+*/
 public class EFGenericRepository<T> : IDisposable, IDataRepository<T> where T : class
 {
-    CareerCloudContext context;
-    DbSet<T> dbSetT;
+    //todo: temporary set to public for testing purposes, should be set back to private
+    public CareerCloudContext context;
+    public DbSet<T> dbSetT;
 
     public EFGenericRepository()
     {
@@ -14,10 +20,49 @@ public class EFGenericRepository<T> : IDisposable, IDataRepository<T> where T : 
         dbSetT = context.Set<T>();
     }
 
+    int CallSaveChanges()
+    {
+        try
+        {
+            var recordCount = context.SaveChanges();
+            if (recordCount == 0)
+                throw new InvalidOperationException("SaveChanges() operation called but 0 rows were affected");
+            return recordCount;
+        }
+
+        //handle errors here
+        catch (ObjectDisposedException) //inherit from InvalidOperationException
+        {
+            throw;
+        }
+        catch (InvalidOperationException)
+        {
+            throw;
+        }
+        catch (DbUpdateConcurrencyException)
+        {
+            throw;
+        }
+
+        catch (DbUpdateException)
+        {
+            throw;
+        }
+
+        //Base EF exception class
+        //catch (DbUpdateException)
+        //{
+        //    throw;
+        //}
+    }
+
     public void Add(params T[] items)
     {
+        if (!items.All(x => x != null))
+            throw new ArgumentNullException(nameof(items));
+
         dbSetT.AddRange(items);
-        context.SaveChanges();
+        CallSaveChanges();
     }
 
     public void CallStoredProc(string name, params Tuple<string, string>[] parameters)
@@ -27,29 +72,73 @@ public class EFGenericRepository<T> : IDisposable, IDataRepository<T> where T : 
 
     public IList<T> GetAll(params Expression<Func<T, object>>[] navigationProperties)
     {
-        throw new NotImplementedException();
+        foreach (var navigationProp in navigationProperties)
+            if (navigationProp is not null)
+                dbSetT.Include(navigationProp);
+
+        //todo: just for test limit top 50 no need tireturn 99999 records
+        var entitiesList = dbSetT
+                            .Take(50)       //used to implement Keyset pagination .where(id >..)
+                            .ToList();
+
+        if (entitiesList is null)
+            throw new InvalidOperationException("No entities found that matches the given predicate.");
+
+        return entitiesList;
     }
 
     public IList<T> GetList(Expression<Func<T, bool>> where, params Expression<Func<T, object>>[] navigationProperties)
     {
-        return dbSetT.Where(where).ToList();
+        if (where is null)
+            throw new ArgumentNullException(nameof(where));
+
+        var linqQueryT = dbSetT.Where(where);
+        foreach (var navigationProp in navigationProperties)
+            if (navigationProp is not null)
+                linqQueryT.Include(navigationProp);
+
+        var entitiesList = linqQueryT.ToList();
+        if (entitiesList is null)
+            throw new InvalidOperationException("No entities found that matches the given predicate.");
+
+        return entitiesList;
     }
 
     public T GetSingle(Expression<Func<T, bool>> where, params Expression<Func<T, object>>[] navigationProperties)
     {
-        return dbSetT.Where(where).FirstOrDefault();
+        if (where is null)
+            throw new ArgumentNullException(nameof(where));
+
+        var linqQueryT = dbSetT.Where(where);
+        foreach (var navigationProp in navigationProperties)
+            if (navigationProp is not null)
+                linqQueryT.Include(navigationProp);
+
+        var entity = linqQueryT.FirstOrDefault();
+        //test is expecting null value to be returned if no record found throwing exception
+        //the method signature should be changed to [T? GetSingle] in the IDataRepository
+
+        return entity;
     }
 
     public void Remove(params T[] items)
     {
+        if (!items.All(x => x != null))
+            throw new ArgumentNullException(nameof(items));
+
+        var dbset = dbSetT.Where(c => items.Contains(c));
         dbSetT.RemoveRange(items);
-        context.SaveChanges();
+        CallSaveChanges();
     }
 
     public void Update(params T[] items)
     {
+        if (!items.All(x => x != null))
+            throw new ArgumentNullException(nameof(items));
+
         dbSetT.UpdateRange(items);
-        context.SaveChanges();
+        CallSaveChanges();
     }
+
     public void Dispose() => context?.Dispose();
 }
